feature/auth added, with login,signup and signout

+++ convex/todos.ts
@@ -1,20 +1,48 @@
+import { Doc } from './_generated/dataModel';
 import {mutation, query} from './_generated/server'
 import { ConvexError, v } from 'convex/values'
 
 
 //mutatations (add,delete, update)
+const isAuthenticated = async  (ctx: any)=>{
+        
+        const identity = await ctx.auth.getUserIdentity();
+        if(identity === null) throw new Error ("Not authenticated")
+        
+        let user = await ctx.db
+            .query("users")
+            .filter((q: { eq: (arg0: any, arg1: any) => any; field: (arg0: string) => any; }) => q.eq(q.field("tokenIdentifier"), identity.tokenIdentifier ))
+            .first();
+        if (user === null) return null
+        return user
+}
+
 export const getTodos = query({
     handler: async (ctx) => {
-        const todos = await ctx.db.query("todos").order("desc").collect()
+        
+        const user = await isAuthenticated(ctx)
+        if (!user)throw new Error ("User not found")
+
+        const todos = await ctx.db.query("todos")
+            .filter((q) => q.eq(q.field("user"), user._id)).order("desc").collect()
+
         return todos
     }
 })
 
+
+
 export const addTodo = mutation({
     args: {text: v.string()},
     handler: async(ctx, args) => {
+
+
+        let user = await isAuthenticated(ctx)
+        if (user === null)throw new Error ("User not found")
+
         const todoId = await ctx.db.insert("todos",{
-            text:args.text,
+            user: user!._id,
+            text: args.text,
             isCompleted: false
         })
         return todoId
@@ -26,8 +54,14 @@ export const addTodo = mutation({
 export const toggleTodo = mutation({
     args: {id:v.id("todos")},
     handler: async(ctx, args) => {
+
+
+        let user = await isAuthenticated(ctx)
+        if (user === null)throw new Error ("User not found")        
+
         const todo  = await ctx.db.get(args.id)
         if (!todo) throw new ConvexError("To do not found")
+        if(todo.user !== user._id) throw new ConvexError("Invalid action")
 
         await ctx.db.patch(args.id, {
             isCompleted: !todo.isCompleted
@@ -42,10 +76,15 @@ export const updateTodo = mutation({
         text:v.string()},
         
     handler: async(ctx, args) => {
+
+        let user = await isAuthenticated(ctx)
+        if (user === null)throw new Error ("User not found")        
+
         const todo  = await ctx.db.get(args.id)
         if (!todo) throw new ConvexError("To do not found")
+        if(todo.user !== user._id) throw new ConvexError("Invalid action")
 
-        await ctx.db.patch(args.id, {
+        await ctx.db.patch((args.id), {
             text:args.text
         })
     }
@@ -54,8 +93,12 @@ export const updateTodo = mutation({
 export const deleteTodo = mutation({
     args: {id:v.id("todos")},
     handler: async(ctx, args) => {
+        let user = await isAuthenticated(ctx)
+        if (user === null)throw new Error ("User not found")        
+
         const todo  = await ctx.db.get(args.id)
         if (!todo) throw new ConvexError("To do not found")
+        if(todo.user !== user._id) throw new ConvexError("Invalid action")
 
         await ctx.db.delete(args.id)
     }
@@ -63,7 +106,10 @@ export const deleteTodo = mutation({
 
 export const deleteAllTodo = mutation({
     handler: async(ctx) => {
-        const todos  = await ctx.db.query("todos").collect()
+        let user = await isAuthenticated(ctx)
+        if (user === null)throw new Error ("User not found")     
+
+        const todos  = await ctx.db.query("todos").filter((q) => q.eq(q.field("user"), user._id)).collect()
 
         for (const todo of todos){
             await ctx.db.delete(todo._id)
